requirements.txt
altgraph==0.17.4
APScheduler==3.11.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
dnspython==2.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpclib==0.4.8
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
multidict==6.5.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pymongo==4.13.2
requests==2.32.4
rsa==4.9.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
viam-sdk==0.48.0
